Refactor type handling for candidateBitsCount variable
Changed candidateBitsCount type from int to byte in the conditional check.
Removed unnecessary cast to byte for validBitsCount since v is already of type byte.
This simplifies the code and ensures proper type handling.

+++ PropertyBitPack.SourceGen/AttributeParsers/BaseAttributeParser.cs
@@ -192,9 +192,9 @@ internal abstract class BaseAttributeParser : IAttributeParser, IContextBindable
         {
             validBitsCount = null;
         }
-        else if(candidateBitsCount is int v)
+        else if(candidateBitsCount is byte v)
         {
-            validBitsCount = (byte)v;
+            validBitsCount = v;
         }
         else
         {
